labirint_test_main/requirements.txt
selenium==4.25.0
pytest==8.3.3
requests==2.32.3
allure-pytest==2.13.5
python-dotenv==1.0.1
pyyaml==6.0.2
